current utils/files/readFiles needs file stats for loading a local fs file. Current configuration load in utils/options did not provide them. Seperated out getFileStats from collectDirectorySize and added it to utils/files exports so that it could be imported in config load. There was also a potential circular dependency, I've reimported the files module in utils/options load config function to ensure it can call readFile when config needs to be loaded
