+++ bids-validator/utils/files/collectDirectorySize.js
@@ -1,4 +1,4 @@
-const fs = require('fs')
+const getFileStats = require('./getFileStats')
 const isNode = require('../isNode')
 
 const collectDirectorySize = fileList => {
@@ -18,18 +18,6 @@ const collectDirectorySize = fileList => {
     }
   })
   return size
-}
-
-const getFileStats = file => {
-  let stats
-  if (!file.stats) {
-    try {
-      stats = fs.statSync(file.path)
-    } catch (err) {
-      stats = { size: 0 }
-    }
-  }
-  return stats
 }
 
 module.exports = collectDirectorySize
