+++ bids-validator/utils/files/getFileStats.js
@@ -1,0 +1,15 @@
+const fs = require('fs')
+
+const getFileStats = file => {
+  let stats
+  if (!file.stats) {
+    try {
+      stats = fs.statSync(file.path)
+    } catch (err) {
+      stats = { size: 0 }
+    }
+  }
+  return stats
+}
+
+module.exports = getFileStats
